# packages/backend/src/generated/swagger.json
{
    "components": {
...
        "schemas": {
            "ApiErrorPayload": {
                "properties": {
                    "error": {
                        "properties": {
                            "data": {
                                "description": "Optional data containing details of the error"
                            },
                            "message": {
                                "type": "string",
                                "description": "A friendly message summarising the error"
                            },
                            "name": {
                                "type": "string",
                                "description": "Unique name for the type of error"
                            },
                            "statusCode": {
                                "type": "number",
                                "format": "integer",
                                "description": "HTTP status code"
                            }
                        },
                        "required": ["name", "statusCode"],
                        "type": "object"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["error"],
                        "nullable": false
                    }
                },
                "required": ["error", "status"],
                "type": "object",
                "description": "The Error object is returned from the api any time there is an error.\nThe message contains"
            },
...
            "ProjectType": {
                "enum": ["DEFAULT", "PREVIEW"],
                "type": "string"
            },
            "OrganizationProject": {
                "properties": {
                    "type": {
                        "$ref": "#/components/schemas/ProjectType"
                    },
                    "name": {
                        "type": "string"
                    },
                    "projectUuid": {
                        "type": "string",
                        "description": "The unique identifier of the project",
                        "format": "uuid"
                    }
                },
                "required": ["type", "name", "projectUuid"],
                "type": "object",
                "description": "Summary of a project under an organization"
            },
            "ApiOrganizationProjects": {
                "properties": {
                    "results": {
                        "items": {
                            "$ref": "#/components/schemas/OrganizationProject"
                        },
                        "type": "array"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["results", "status"],
                "type": "object",
                "description": "List of projects in the current organization"
            },
...
            "DbtLocalProjectConfig": {
                "properties": {
                    "type": {
                        "$ref": "#/components/schemas/DbtProjectType.DBT"
                    },
                    "target": {
                        "type": "string"
                    },
                    "environment": {
                        "items": {
                            "$ref": "#/components/schemas/DbtProjectEnvironmentVariable"
                        },
                        "type": "array"
                    },
                    "profiles_dir": {
                        "type": "string"
                    },
                    "project_dir": {
                        "type": "string"
                    }
                },
                "required": ["type"],
                "type": "object",
                "additionalProperties": false
            },
...
            "DbtCloudIDEProjectConfig": {
                "properties": {
                    "type": {
                        "$ref": "#/components/schemas/DbtProjectType.DBT_CLOUD_IDE"
                    },
                    "api_key": {
                        "type": "string"
                    },
                    "account_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "number",
                                "format": "double"
                            }
                        ]
                    },
                    "environment_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "number",
                                "format": "double"
                            }
                        ]
                    },
                    "project_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "number",
                                "format": "double"
                            }
                        ]
                    }
                },
                "required": [
                    "type",
                    "api_key",
                    "account_id",
                    "environment_id",
                    "project_id"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
            "DbtGithubProjectConfig": {
                "properties": {
                    "type": {
                        "$ref": "#/components/schemas/DbtProjectType.GITHUB"
                    },
                    "target": {
                        "type": "string"
                    },
                    "environment": {
                        "items": {
                            "$ref": "#/components/schemas/DbtProjectEnvironmentVariable"
                        },
                        "type": "array"
                    },
                    "personal_access_token": {
                        "type": "string"
                    },
                    "repository": {
                        "type": "string"
                    },
                    "branch": {
                        "type": "string"
                    },
                    "project_sub_path": {
                        "type": "string"
                    },
                    "host_domain": {
                        "type": "string"
                    }
                },
                "required": [
                    "type",
                    "personal_access_token",
                    "repository",
                    "branch",
                    "project_sub_path"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
            "DbtNoneProjectConfig": {
                "properties": {
                    "type": {
                        "$ref": "#/components/schemas/DbtProjectType.NONE"
                    },
                    "target": {
                        "type": "string"
                    },
                    "environment": {
                        "items": {
                            "$ref": "#/components/schemas/DbtProjectEnvironmentVariable"
                        },
                        "type": "array"
                    },
                    "hideRefreshButton": {
                        "type": "boolean"
                    }
                },
                "required": ["type"],
                "type": "object",
                "additionalProperties": false
            },
            "DbtProjectConfig": {
                "anyOf": [
                    {
                        "$ref": "#/components/schemas/DbtLocalProjectConfig"
                    },
                    {
                        "$ref": "#/components/schemas/DbtCloudIDEProjectConfig"
                    },
                    {
                        "$ref": "#/components/schemas/DbtGithubProjectConfig"
                    },
                    {
                        "$ref": "#/components/schemas/DbtBitBucketProjectConfig"
                    },
                    {
                        "$ref": "#/components/schemas/DbtGitlabProjectConfig"
                    },
                    {
                        "$ref": "#/components/schemas/DbtAzureDevOpsProjectConfig"
                    },
                    {
                        "$ref": "#/components/schemas/DbtNoneProjectConfig"
                    }
                ]
            },
...
            "SnowflakeCredentials": {
                "$ref": "#/components/schemas/Omit_CreateSnowflakeCredentials.SensitiveCredentialsFieldNames_"
            },
...
            "RedshiftCredentials": {
                "$ref": "#/components/schemas/Omit_CreateRedshiftCredentials.SensitiveCredentialsFieldNames_"
            },
...
            "PostgresCredentials": {
                "$ref": "#/components/schemas/Omit_CreatePostgresCredentials.SensitiveCredentialsFieldNames_"
            },
...
            "BigqueryCredentials": {
                "$ref": "#/components/schemas/Omit_CreateBigqueryCredentials.SensitiveCredentialsFieldNames_"
            },
...
            "DatabricksCredentials": {
                "$ref": "#/components/schemas/Omit_CreateDatabricksCredentials.SensitiveCredentialsFieldNames_"
            },
...
            "TrinoCredentials": {
                "$ref": "#/components/schemas/Omit_CreateTrinoCredentials.SensitiveCredentialsFieldNames_"
            },
            "WarehouseCredentials": {
                "anyOf": [
                    {
                        "$ref": "#/components/schemas/SnowflakeCredentials"
                    },
                    {
                        "$ref": "#/components/schemas/RedshiftCredentials"
                    },
                    {
                        "$ref": "#/components/schemas/PostgresCredentials"
                    },
                    {
                        "$ref": "#/components/schemas/BigqueryCredentials"
                    },
                    {
                        "$ref": "#/components/schemas/DatabricksCredentials"
                    },
                    {
                        "$ref": "#/components/schemas/TrinoCredentials"
                    }
                ]
            },
            "SupportedDbtVersions": {
                "enum": ["v1.4", "v1.5"],
                "type": "string"
            },
            "Project": {
                "properties": {
                    "dbtVersion": {
                        "$ref": "#/components/schemas/SupportedDbtVersions"
                    },
                    "copiedFromProjectUuid": {
                        "type": "string"
                    },
                    "pinnedListUuid": {
                        "type": "string"
                    },
                    "warehouseConnection": {
                        "$ref": "#/components/schemas/WarehouseCredentials"
                    },
                    "dbtConnection": {
                        "$ref": "#/components/schemas/DbtProjectConfig"
                    },
                    "type": {
                        "$ref": "#/components/schemas/ProjectType"
                    },
                    "name": {
                        "type": "string"
                    },
                    "projectUuid": {
                        "type": "string"
                    },
                    "organizationUuid": {
                        "type": "string"
                    }
                },
                "required": [
                    "dbtVersion",
                    "dbtConnection",
                    "type",
                    "name",
                    "projectUuid",
                    "organizationUuid"
                ],
                "type": "object"
            },
            "ApiProjectResponse": {
                "properties": {
                    "results": {
                        "$ref": "#/components/schemas/Project"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["results", "status"],
                "type": "object"
            },
...
            "ApiJobScheduledResponse": {
                "properties": {
                    "results": {
                        "properties": {
                            "jobId": {
                                "type": "string"
                            }
                        },
                        "required": ["jobId"],
                        "type": "object"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["results", "status"],
                "type": "object"
            },
...
            "ValidationErrorChartResponse": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ValidationResponseBase"
                    },
                    {
                        "properties": {
                            "chartName": {
                                "type": "string"
                            },
                            "chartViews": {
                                "type": "number",
                                "format": "double"
                            },
                            "lastUpdatedAt": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "lastUpdatedBy": {
                                "type": "string"
                            },
                            "fieldName": {
                                "type": "string"
                            },
                            "chartType": {
                                "$ref": "#/components/schemas/ChartKind"
                            },
                            "chartUuid": {
                                "type": "string"
                            }
                        },
                        "required": ["chartViews"],
                        "type": "object"
                    }
                ]
            },
            "ValidationErrorDashboardResponse": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ValidationResponseBase"
                    },
                    {
                        "properties": {
                            "dashboardViews": {
                                "type": "number",
                                "format": "double"
                            },
                            "lastUpdatedAt": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "lastUpdatedBy": {
                                "type": "string"
                            },
                            "fieldName": {
                                "type": "string"
                            },
                            "chartName": {
                                "type": "string"
                            },
                            "dashboardUuid": {
                                "type": "string"
                            }
                        },
                        "required": ["dashboardViews"],
                        "type": "object"
                    }
                ]
            },
...
            "Omit_ValidationResponseBase.name_": {
                "$ref": "#/components/schemas/Pick_ValidationResponseBase.Exclude_keyofValidationResponseBase.name__",
                "description": "Construct a type with the properties of T except for those in type K."
            },
            "ValidationErrorTableResponse": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Omit_ValidationResponseBase.name_"
                    },
                    {
                        "properties": {
                            "name": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "ValidationResponse": {
                "anyOf": [
                    {
                        "$ref": "#/components/schemas/ValidationErrorChartResponse"
                    },
                    {
                        "$ref": "#/components/schemas/ValidationErrorDashboardResponse"
                    },
                    {
                        "$ref": "#/components/schemas/ValidationErrorTableResponse"
                    }
                ]
            },
            "ApiValidateResponse": {
                "properties": {
                    "results": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationResponse"
                        },
                        "type": "array"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["results", "status"],
                "type": "object"
            },
            "ApiValidationDismissResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["status"],
                "type": "object"
            }
        },
        "securitySchemes": {
            "session_cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "connect.sid"
            },
            "api_key": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Value should be 'ApiKey <your key>'"
            }
        }
    },
    "info": {
        "title": "Lightdash API",
        "version": "0.714.0",
        "description": "Open API documentation for all public Lightdash API endpoints",
        "license": {
            "name": "MIT"
        },
        "contact": {
            "name": "Lightdash Support",
            "email": "support@lightdash.com",
            "url": "https://docs.lightdash.com/help-and-contact/contact/contact_info/"
        }
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/org/projects": {
            "get": {
                "operationId": "ListOrganizationProjects",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiOrganizationProjects"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Gets all projects of the current user's organization",
                "tags": ["Organizations"],
                "security": [],
                "parameters": []
            }
        },
...
        "/api/v1/projects/{projectUuid}": {
            "get": {
                "operationId": "GetProject",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiProjectResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Get a project of an organiztion",
                "tags": ["Projects"],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "projectUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
        "/api/v1/projects/{projectUuid}/validate": {
            "post": {
                "operationId": "ValidateProject",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiJobScheduledResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Validate content inside a project. This will start a validation job and return the job id.\n\nValidation jobs scan all charts and dashboards inside a project to find any broken references\nto metrics or dimensions that aren't available. Results are available after the job is completed.",
                "tags": ["Projects"],
                "security": [],
                "parameters": [
                    {
                        "description": "the projectId for the validation",
                        "in": "path",
                        "name": "projectUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "the compiled explores to validate against an existing project, this is used in the CLI to validate a project without creating a preview",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "explores": {
                                        "items": {},
                                        "type": "array"
                                    }
                                },
                                "type": "object",
                                "description": "the compiled explores to validate against an existing project, this is used in the CLI to validate a project without creating a preview"
                            }
                        }
                    }
                }
            },
            "get": {
                "operationId": "GetLatestValidationResults",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiValidateResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Get validation results for a project. This will return the results of the latest validation job.",
                "tags": ["Projects"],
                "security": [],
                "parameters": [
                    {
                        "description": "the projectId for the validation",
                        "in": "path",
                        "name": "projectUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "boolean to know if this request is made from the settings page, for analytics",
                        "in": "query",
                        "name": "fromSettings",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "description": "optional jobId to get results for a specific job, used on CLI",
                        "in": "query",
                        "name": "jobId",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/v1/projects/{projectUuid}/validate/{validationId}": {
            "delete": {
                "operationId": "DeleteValidationDismiss",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiValidationDismissResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Deletes a single validation error.",
                "tags": ["Projects"],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "projectUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "the projectId for the validation",
                        "in": "path",
                        "name": "validationId",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ]
            }
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ],
...
}